requirements.txt
fastapi==0.75.1
uvicorn==0.17.6
pydantic==1.9.1
pika==1.2.0
sqlalchemy==1.4.35
databases==0.5.5
alembic==1.7.6
python-dotenv==0.20.0
pytest
requests